+++ tybotRoute/tiledeskChatbotPlugs/DirectivesChatbotPlug.js
@@ -458,7 +458,7 @@ class DirectivesChatbotPlug {
       });
     }
     else if (directive_name === Directives.SEND_EMAIL) {
-      console.log("...DirSendEmail");
+      // console.log("...DirSendEmail");
       new DirSendEmail(context).execute(directive, async () => {
         let next_dir = await this.nextDirective(this.directives);
         this.process(next_dir);